Logger Project/python/requirements.txt
bidict==0.23.1
certifi==2024.2.2
charset-normalizer==3.3.2
dnspython==2.6.1
eventlet==0.36.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
numpy==1.26.4
pandas==2.2.2
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2024.1
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
wsproto==1.2.0
